Extracting more code from the plugin
