+++ docs/api/internals/store.rst
@@ -3,7 +3,7 @@ The Store
 
 .. autoclass:: extended_mypy_django_plugin.plugin._store.Store
 
-.. autoprotocol:: extended_mypy_django_plugin.plugin._store.LookupFunction
+.. autoprotocol:: extended_mypy_django_plugin.plugin._store.LookupInfo
 
 .. autoprotocol:: extended_mypy_django_plugin.plugin._store.LookupInstanceFunction
