+++ extended_mypy_django_plugin/plugin/actions/__init__.py
@@ -1,5 +1,12 @@
 from ._annotation_resolver import AnnotationResolver
 from ._sem_analyze import SemAnalyzing, TypeAnalyzer
-from ._type_checker import TypeChecking
+from ._type_checker import SharedAnnotationHookLogic, SharedSignatureHookLogic, TypeChecking
 
-__all__ = ["SemAnalyzing", "TypeAnalyzer", "AnnotationResolver", "TypeChecking"]
+__all__ = [
+    "SemAnalyzing",
+    "TypeAnalyzer",
+    "AnnotationResolver",
+    "TypeChecking",
+    "SharedAnnotationHookLogic",
+    "SharedSignatureHookLogic",
+]
